split letter.js into components
