+++ src/admin/dumb_components/letter_header.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export default () => {
+  return (
+    <div className='header'>
+      <h1>Friends of Chichester Harbour</h1>
+      <h3>Registered Charity No: 1051162</h3>
+      <img src='/images/logo.png' className='letter-logo'></img>
+    </div>
+  )
+}
